refactor conrollers
